Revert "Atualizando para versão mais recente"
This reverts commit aa2e74e1e3fde4485151954cb6b7721d3a03ba7a.

+++ DesafioCliente/Models/Cliente.cs
@@ -9,10 +9,10 @@ namespace ClienteManagement.Models
         public int ClienteId { get; set; }
         public string ClienteName { get; set; } = string.Empty;
         public string ClienteLastName { get; set; } = string.Empty;
-        public long ClientePhone { get; set; }
+        public int ClientePhone { get; set; }
         public string ClienteEmail { get; set; } = string.Empty;
         public string ClienteCep { get; set; } = string.Empty;
-        public long ClienteCPF { get; set; }
+        public int ClienteCPF { get; set; }
     }
 
 }
